windows: added support to skip building externals with github using a special cache file

+++ vs/extern/dragonscript/build.ps1
@@ -8,6 +8,11 @@
 Import-Module "$PSScriptRoot\..\..\shared.psm1"
 
 
+if (Test-Path "$PSScriptRoot\..\..\github_cached_externals") {
+    return
+}
+
+
 $ExpandedDir = "$ProjectDir\dragonscript-1.5.1"
 if (Test-Path $ExpandedDir) {
     Remove-Item $ExpandedDir -Force -Recurse
