+++ vs/extern/libwebm/build.ps1
@@ -8,6 +8,11 @@
 Import-Module "$PSScriptRoot\..\..\shared.psm1"
 
 
+if (Test-Path "$PSScriptRoot\..\..\github_cached_externals") {
+    return
+}
+
+
 $ExpandedDir = Join-Path -Path $ProjectDir -ChildPath "libwebm-libwebm-1.0.0.29-vc64"
 if (Test-Path $ExpandedDir) {
     Remove-Item $ExpandedDir -Force -Recurse
